Refactor to make TracingContext interface for holding  span stack
This allows us to be more flexible in how we track related spans, especially for systems that
use fixed-size thread pools that recycle threads (which can cause problems with ThreadLocals
because they can inherit incorrect contexts, etc).

Also, making some small adjustments to Span.toString() to allow it to report on actual
class / sub-class in use.