fix: set service name in spring boot starter (#317)
## Which problem is this PR solving?

- When using a non-classic API key... `service.name` is not properly
set, resulting in traces ending up in wrong dataset. For example, with
`dataset=my-dataset` and `service.name=my-service`, traces should show
up in `my-service` dataset.

## Short description of the changes

- set servicename in spring boot starter
- use process name instead of service name for default fallback
- don't throw on runtime exception
- use dataset for classic keys and service name for nonclassic keys
- update and add unit tests to assert ServiceName is properly used for
dataset with nonclassic keys, and dataset for classic keys
- update fallback `service.name` attribute to be `unknown_service:java`.

Classic Key: set `honeycomb.beeline.dataset` and that is where data ends
up in Honeycomb
NonClassic Key: set `honeycomb.beeline.service-name` and that is where
data ends up in Honeycomb

**NOTE**: this is technically breaking. Data was never being sent to
`service.name` dataset, so it doesn't change location in Honeycomb for
classic keys. If using a Classic key, and the `service.name` attribute
was left empty and relied on `spring.application.name`, it will now have
`unknown_service:java` for that attribute but will continue to send to
the configured dataset. If using a nonclassic key (any new user), data
will be sent to the configured `service-name` attribute; if that is not
set, it will be sent to `unknown_service:java` in Honeycomb. This should
be called out in the release notes